// src/lib/enums.ts
export enum PropertyType {
	Apartment = "Apartment",
	House = "House",
	Cabin = "Cabin",
	Villa = "Villa",
	Cottage = "Cottage",
	Bungalow = "Bungalow",
	Loft = "Loft",
	Studio = "Studio",
	Townhouse = "Townhouse",
	Duplex = "Duplex",
	Other = "Other",
}

export enum PropertyStatus {
	Inactive = "Inactive",
	Pending = "Pending",
	Active = "Active",
}